overlord/state: move state lock tracing, unexport functions, change build tag
Move state lock tracing from osutil into overlord/state. Tweak the code
to not export functions since there aren't any external consumers.
